requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
Flask-WTF==1.2.1
Flask-Limiter==3.8.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
